requirements.txt
python-dotenv
streamlit
PyPDF2
langchain
openai
tiktoken
faiss-cpu
